separating to utils
